twitterimagecreater: the Side: guards test unsigned, not signed
All nine ...Side: setters bound their column index with
"if (side < kScoreColumnCount)" over a signed int. The binary's guard is
cmp #1 followed by b.ls, or b.hi to skip at 0x87a70, and both are unsigned.

The two agree for every sensible input, which is why it survived review. On
side = -1 the reconstruction passes and evaluates m_Score[-1], reading an
object pointer from before the array, while the binary rejects it because
(unsigned)-1 is far above 1.

Each of the nine guards was read individually rather than the family inferred
from one, since a single member using a signed compare would have made a
blanket change wrong. All nine now cast to unsigned int.

The condition codes are the tell and it generalises: b.lt and b.ge are
signed, b.ls and b.hi are not, and the decompiler renders both as "<". An
array index guarded only from above is worth re-reading for it.

Also verifies AVBus -removeSource and -audioPlayerBeginInterruption:, and
RBMusicDifficultyView -getDifficultyButton:.

+++ Project/TwitterImageCreater.mm
@@ -294,63 +294,63 @@ static const CGPoint g_TwitterFullComboPos = {25.0, 135.0};
 
 - (void)setScore:(int)score Side:(int)side {
     /** @ghidraAddress 0x87930 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setScore:score];
     }
 }
 
 - (void)setAR:(float)aR Side:(int)side {
     /** @ghidraAddress 0x87958 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setAr:aR];
     }
 }
 
 - (void)setJustNum:(int)justNum Side:(int)side {
     /** @ghidraAddress 0x87980 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setJustNum:justNum];
     }
 }
 
 - (void)setGreatNum:(int)greatNum Side:(int)side {
     /** @ghidraAddress 0x879a8 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setGreatNum:greatNum];
     }
 }
 
 - (void)setGoodNum:(int)goodNum Side:(int)side {
     /** @ghidraAddress 0x879d0 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setGoodNum:goodNum];
     }
 }
 
 - (void)setMissNum:(int)missNum Side:(int)side {
     /** @ghidraAddress 0x879f8 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setMissNum:missNum];
     }
 }
 
 - (void)setJustReflecNum:(int)justReflecNum Side:(int)side {
     /** @ghidraAddress 0x87a20 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setJustReflecNum:justReflecNum];
     }
 }
 
 - (void)setMaxComboNum:(int)maxComboNum Side:(int)side {
     /** @ghidraAddress 0x87a48 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setMaxComboNum:maxComboNum];
     }
 }
 
 - (void)setName:(NSString *)name Side:(int)side {
     /** @ghidraAddress 0x87a70 */
-    if (side < kScoreColumnCount) {
+    if (static_cast<unsigned int>(side) < kScoreColumnCount) {
         [m_Score[side] setName:name];
     }
 }
